I hate this code
