+++ glances/plugins/plugin/model.py
@@ -14,6 +14,8 @@ I am your father...
 
 import copy
 import re
+import time
+
 
 from glances.actions import GlancesActions
 from glances.events_list import glances_events
@@ -22,7 +24,9 @@ from glances.history import GlancesHistory
 from glances.logger import logger
 from glances.outputs.glances_unicode import unicode_message
 from glances.thresholds import glances_thresholds
-from glances.timer import Counter, Timer, getTimeSinceLastUpdate, debounce
+from glances.timer import Counter, Timer, getTimeSinceLastUpdate
+from threading import Timer
+from inspect import signature
 
 fields_unit_short = {'percent': '%'}
 
@@ -494,6 +498,42 @@ class GlancesPluginModel:
         self.views = ret
 
         return self.views
+
+    def debounce(wait):
+        def decorator(fn):
+            sig = signature(fn)
+            caller = {}
+
+            def debounced(*args, **kwargs):
+                nonlocal caller
+
+                try:
+                    bound_args = sig.bind(*args, **kwargs)
+                    bound_args.apply_defaults()
+                    called_args = fn.__name__ + str(dict(bound_args.arguments))
+                except:
+                    called_args = ''
+
+                def call_it(key):
+                    try:
+                        # always remove on call
+                        caller.pop(key)
+                    except:
+                        pass
+
+                    fn(*args, **kwargs)
+
+                try:
+                    # Always try to cancel timer
+                    caller[called_args].cancel()
+                except:
+                    pass
+
+                caller[called_args] = Timer(wait, call_it, [called_args])
+                caller[called_args].start()
+
+            return debounced
+        return decorator
 
     @debounce(0.3) 
     def manageElseIf(self, value, field, ret):
